Allow for compilation of app specific Sass
