+++ ecommerce/adyen/templates/oscar/basket/partials/basket_content.html
@@ -168,7 +168,7 @@
 
         <div class="col-md-6 col-xs-12 border-left-md payment-details">
 
-            <form method="POST" action="/payment/adyen/payment/" id="adyen-encrypted-form">
+            <form method="POST" action="{% url 'adyen:payment' %}" id="adyen-encrypted-form">
                 {% for processor in payment_processors %}
                     {% if processor.NAME|lower == 'adyen'%}
                         <input type="hidden" value="{{ processor.generation_time }}" data-encrypted-name="generationtime" />
