+++ ecommerce/adyen/urls.py
@@ -1,0 +1,9 @@
+""" Payment-related URLs """
+from django.conf.urls import url
+
+from ecommerce.extensions.payment import views
+
+urlpatterns = [
+    url(r'^notification/$', views.AdyenNotificationView.as_view(), name='notification'),
+    url(r'^payment/$', views.AdyenPaymentView.as_view(), name='payment'),
+]
